Add files via upload
[the hotel app]
12/16/25 last modified via MacBook

introduction:
An ongoing major project over the last 5 months for our freshman year class at the university. It is a joint collaboration between my classmate and me for an Object-Oriented Programming (OOP) class.

description:
This is a GUI-themed application using the Tkinter framework built on PyCharm for creating our buttons & widgets. Essentially, in our class project, we will be presenting this as a demo of what we've been working on over the first half of the semester, on 12/17/2025.

footnotes:
gui scripting, application, python, coding, database framework, pickle

github developers:
zodeelee, Xander10Alegria